cleaned up contribution model framework so that everything fits into a driver, plugin, or core
