+++ dsi/drivers/filesystem/tests/query.csv
@@ -1,0 +1,3 @@
+
+2,180,0.05,ST5_FF_DUET,ST5_ignite_contour.dat,safe,safe,yes,,122388,https://wifire-data.sdsc.edu/data//burnpro3d/d/07/12/run_07127fcc-66c4-4498-a4c6-8e9f6e5fbe68/png/run_07127fcc-66c4-4498-a4c6-8e9f6e5fbe68_fuels-dens_2100_000.png,2
+8,225,0.05,ST5_FFDUET_TLS,ST5_ignite_contour.dat,safe,unsafe,no,,169302,https://wifire-data.sdsc.edu/data//burnpro3d/d/cd/34/run_cd34e198-92d8-465d-b7d7-9cdf8b4dbd12/png/run_cd34e198-92d8-465d-b7d7-9cdf8b4dbd12_fuels-dens_2100_000.png,8
